pyproject.toml
[project]
name = "lets-not-go-bankrupt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "camelot-py[cv]>=1.0.0",
    "ghostscript>=0.7",
    "jpype1>=1.5.2",
    "pandas>=2.2.3",
    "pandas-gbq>=0.28.1",
    "polars>=1.29.0",
    "pydantic>=2.11.4",
    "pydantic-ai[openai]>=0.2.4",
    "pydantic-ai-slim[openai]>=0.2.4",
    "pypdf2>=3.0.1",
    "python-telegram-bot>=22.1",
    "tabula-py>=2.10.0",
]
